+++ server/test/e2e/surrender.e2e-spec.ts
@@ -1,0 +1,122 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { INestApplication } from '@nestjs/common';
+import * as request from 'supertest';
+import LEVEL from '../../src/domain/constant/LEVEL';
+import { AppModule } from '../../src/app.module';
+import { DataServicesModule } from '../../src/frameworks/data-services/DataService.module';
+import { GungiRepository } from '../../src/frameworks/data-services/GungiRepository';
+import GungiDataModel from '../../src/frameworks/data-services/data-model/GungiDataModel';
+import { GungiData } from '../../src/frameworks/data-services/GungiData';
+import { randomUUID } from 'crypto';
+import SIDE from '../../src/domain/constant/SIDE';
+import { Db, MongoClient } from 'mongodb';
+import * as dotenv from 'dotenv';
+import { ZodFilter } from '../../src/frameworks/filter/ZodFilter';
+import { APP_FILTER } from '@nestjs/core';
+
+dotenv.config();
+
+describe('surrender (e2e)', () => {
+  let app: INestApplication;
+  let gungiRepository: GungiRepository;
+  let gungiDataModel: GungiDataModel;
+  let db: Db;
+  let client: MongoClient;
+
+  beforeEach(async () => {
+    const moduleFixture: TestingModule = await Test.createTestingModule({
+      imports: [AppModule, DataServicesModule],
+      providers: [
+        GungiDataModel,
+        {
+          provide: APP_FILTER,
+          useClass: ZodFilter,
+        },
+      ],
+    }).compile();
+
+    app = moduleFixture.createNestApplication();
+    gungiDataModel = moduleFixture.get<GungiDataModel>(GungiDataModel);
+    gungiRepository = moduleFixture.get('GungiRepository');
+    client = moduleFixture.get('MongoConnection');
+    db = client.db();
+
+    await app.init();
+  });
+
+  afterEach(async () => {
+    await db.collection('Gungi').deleteMany({});
+    await client.close(true);
+    await app.close();
+  });
+
+  it('/(POST) gungi/:gungiId/surrender', async () => {
+    const gungiId = randomUUID();
+    const gungiData: GungiData = {
+      currentTurn: SIDE.WHITE,
+      gungiHan: { han: [] },
+      history: [],
+      _id: gungiId,
+      level: LEVEL.BEGINNER,
+      turn: {
+        sente: 'A',
+        gote: 'B',
+      },
+      players: [
+        {
+          id: 'A',
+          name: 'A',
+          side: SIDE.WHITE,
+          gomaOki: { gomas: [] },
+          deadArea: { gomas: [] },
+        },
+        {
+          id: 'B',
+          name: 'B',
+          side: SIDE.BLACK,
+          gomaOki: { gomas: [] },
+          deadArea: { gomas: [] },
+        },
+      ],
+    };
+
+    const gungi = gungiDataModel.toDomain(gungiData);
+
+    await gungiRepository.save(gungi);
+
+    // post a  json
+    const body = {
+      playerId: 'A',
+    };
+
+    await request(app.getHttpServer())
+      .post(`/gungi/${gungiId}/surrender`)
+      .send(body)
+      .expect(200)
+      .expect({ winner: 'B' });
+  });
+
+  it('should return 400 when request body is empty', async () => {
+    const body = {};
+    await request(app.getHttpServer())
+      .post('/gungi/someGungiId/surrender')
+      .send(body)
+      .expect(400);
+  });
+
+  it('should return 400 when "playerId" field is missing', async () => {
+    const body = { someOtherField: 'value' };
+    await request(app.getHttpServer())
+      .post('/gungi/someGungiId/surrender')
+      .send(body)
+      .expect(400);
+  });
+
+  it('should return 400 when "playerId" field is an empty string', async () => {
+    const body = { playerId: '' };
+    await request(app.getHttpServer())
+      .post('/gungi/someGungiId/surrender')
+      .send(body)
+      .expect(400);
+  });
+});
